Use thiserror for decode error handling
ResultStr has since the beginning been a trivial placeholder type until
real error codes could be instituted. This change starts (and nearly
completes) the error code conversion of the codebase. This is a breaking
change for the greater good.

While working on this I also changed the signature of some of the
decoding functions to instead return a Vec of errors. That means e.g.
AVP errors should now actually be propagated to the end user.